ENH: Add new parameter to schedule_function that accepts a trading calendar.
